+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [2.0.2] 7th August 2025
 - Introduced preview in 'Advanced Web-popup Builder' in 'Web Popup'.
+- - Added support for subdomain specific storage
+
 
 ## [2.0.1] 7th August 2025
 - Corrected click actions for web popup image only template